pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "toolbox"
version = "2.0.0"
requires-python = ">=3.10"
description = "Helper tools for MRI development and reconstruction workflows"

authors = [
    { name = "Jinho Kim", email = "jinho.kim@fau.de" }
]

license = { file = "LICENSE" }

keywords = ["MRI", "reconstruction", "image-processing", "research"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Topic :: Scientific/Engineering",
    "Topic :: Scientific/Engineering :: Image Processing",
]

dependencies = [
    "cupy-cuda11x==13.2.0",
    "h5py==3.11.0",
    "jupyter==1.0.0",
    "matplotlib==3.9.0",
    "numpy==1.24.4",
    "pygrappa==0.26.2",
    "scikit-image==0.23.2",
    "scikit-learn==1.6.1",
    "scipy==1.13.1",
    "sigpy==0.1.26",
]

[project.urls]
Homepage = "https://github.com/JinhoKim46/mri_toolbox"

[tool.setuptools]
packages = ["toolbox"]
